+++ lib/app/presentation/user/views/login.dart
@@ -1,0 +1,41 @@
+import 'dart:developer';
+import 'package:petmeals/app/presentation/user/widgets/btn.login.dart';
+import 'package:petmeals/app/presentation/user/widgets/intro.login.dart';
+import 'package:petmeals/app/presentation/user/widgets/logo.login.dart';
+import 'package:flutter/material.dart';
+
+class LoginView extends StatelessWidget {
+  const LoginView({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    log('login *');
+    return Scaffold(
+      body: SafeArea(
+        child: Container(
+          margin: const EdgeInsets.symmetric(horizontal: 20),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.center,
+            children: [
+              Expanded(
+                child: Container(
+                  margin: const EdgeInsets.only(top: 25, bottom: 0),
+                  width: double.maxFinite,
+                  child: ListView(
+                    children: const [
+                      LogoWidget(),
+                      SizedBox(height: 8.5),
+                      IntroWidget(),
+                      SizedBox(height: 32),
+                      ButtonGoogle(),
+                    ],
+                  ),
+                ),
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
