--- lib/config/utils/data/breeds.list.dart
@@ -1,2794 +1,0 @@
-import 'package:petmeals/config/utils/model/breed.model.dart';
-
-final breedList = [
-  BreedModel(
-    breedId: 1,
-    breedName: 'Affenpinscher',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 2,
-    breedName: 'Aïdi - Perro del Atlas',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 3,
-    breedName: 'Airedale Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 4,
-    breedName: 'Akita Americano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 5,
-    breedName: 'Akita Inu',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 6,
-    breedName: 'Alano Español',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 7,
-    breedName: 'Alaskan Malamute',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 8,
-    breedName: 'American Bully',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 9,
-    breedName: 'American Eskimo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 10,
-    breedName: 'American Pit Bull Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 11,
-    breedName: 'American Staffordshire-Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 12,
-    breedName: 'Anglo-Français de Petite Vénerie',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 13,
-    breedName: 'Antiguo Perro de Muestra Danés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 14,
-    breedName: 'Antiguo Perro de Pastor Inglés - Bobtail',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 15,
-    breedName: 'Australian Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 16,
-    breedName: 'Azawakh',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 17,
-    breedName: 'Barbet',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 18,
-    breedName: 'Basenji',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 19,
-    breedName: 'Basset Artésien-Normand',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 20,
-    breedName: 'Basset Azul de Gascuña',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 21,
-    breedName: 'Basset Fauve de Bretagne',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 22,
-    breedName: 'Basset-Hound',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 23,
-    breedName: 'Beagle',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 24,
-    breedName: 'Beagle Harrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 25,
-    breedName: 'Bearded Collie',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 26,
-    breedName: 'Bedlington Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 27,
-    breedName: 'Bichón Boloñés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 28,
-    breedName: 'Bichón Frisé',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 29,
-    breedName: 'Bichón Habanero',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 30,
-    breedName: 'Bichón Maltés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 31,
-    breedName: 'Biewer Yorkshire Terrier a la Pom Pon',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 32,
-    breedName: 'Billy',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 33,
-    breedName: 'Black and Tan Coonhound',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 34,
-    breedName: 'Bluetick Coonhound',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 35,
-    breedName: 'Boerboel',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 36,
-    breedName: 'Border Collie',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 37,
-    breedName: 'Border Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 38,
-    breedName: 'Borzoi',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 39,
-    breedName: 'Boston Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 40,
-    breedName: 'Bouvier des Ardennes',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 41,
-    breedName: 'Boxer',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 42,
-    breedName: 'Boyero de Appenzell',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 43,
-    breedName: 'Boyero de Australia',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 44,
-    breedName: 'Boyero de Berna',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 45,
-    breedName: 'Boyero de Entlebuch',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 46,
-    breedName: 'Boyero de Flandes',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 47,
-    breedName: 'Braco Alemán de Pelo Corto',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 48,
-    breedName: 'Braco Alemán de Pelo Duro',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 49,
-    breedName: 'Braco de Auvernia',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 50,
-    breedName: 'Braco del Ariège',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 51,
-    breedName: 'Braco del Bourbonnais',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 52,
-    breedName: 'Braco Francés tipo Gascuña Talla Grande',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 53,
-    breedName: 'Braco Francés tipo Pirineos Talla Pequeña',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 54,
-    breedName: 'Braco Húngaro de Pelo Corto',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 55,
-    breedName: 'Braco Húngaro de Pelo Duro',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 56,
-    breedName: 'Braco Italiano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 57,
-    breedName: 'Braco San Germain',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 58,
-    breedName: 'Briquet Griffon Vendéen',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 59,
-    breedName: 'Broholmer',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 60,
-    breedName: 'Bull Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 61,
-    breedName: 'Bull Terrier Miniatura',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 62,
-    breedName: 'Bulldog Americano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 63,
-    breedName: 'Bulldog Francés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 64,
-    breedName: 'Bulldog Inglés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 65,
-    breedName: 'Bullmastiff',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 66,
-    breedName: 'Ca de Bestiar - Perro de Pastor Mallorquín',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 67,
-    breedName: 'Ca de Bou - Perro de Presa Mallorquín',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 68,
-    breedName: 'Ca Rater Mallorquí',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 69,
-    breedName: 'Cairn Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 70,
-    breedName: 'Can de Palleiro',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 71,
-    breedName: 'Can Guicho ou Quisquelo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 72,
-    breedName: 'Canaan Dog',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 73,
-    breedName: 'Cane Corso',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 74,
-    breedName: 'Caniche Enano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 75,
-    breedName: 'Caniche Gigante',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 76,
-    breedName: 'Caniche Mediano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 77,
-    breedName: 'Caniche Toy',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 78,
-    breedName: 'Cão da Serra da Estrela',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 79,
-    breedName: 'Cão da Serra de Aires',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 80,
-    breedName: 'Cão de Castro Laboreiro',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 81,
-    breedName: 'Carlino - Pug',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 82,
-    breedName: 'Catahoula Leopard Dog',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 83,
-    breedName: 'Cavalier King Charles Spaniel',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 84,
-    breedName: 'Cesky Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 85,
-    breedName: 'Chesapeake Bay Retriever',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 86,
-    breedName: 'Chien d Artois',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 87,
-    breedName: 'Chihuahua',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 88,
-    breedName: 'Chinook',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 89,
-    breedName: 'Chow Chow',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 90,
-    breedName: 'Cimarrón Uruguayo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 91,
-    breedName: 'Cirneco del Etna',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 92,
-    breedName: 'Clumber Spaniel',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 93,
-    breedName: 'Cocker Spaniel Americano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 94,
-    breedName: 'Cocker Spaniel Inglés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 95,
-    breedName: 'Collie de Pelo Corto',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 96,
-    breedName: 'Collie de Pelo Largo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 97,
-    breedName: 'Coton de Tuléar',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 98,
-    breedName: 'Dálmata',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 99,
-    breedName: 'Dandie Dinmont Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 100,
-    breedName: 'Deerhound',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 101,
-    breedName: 'Dobermann',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 102,
-    breedName: 'Dogo Alemán',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 103,
-    breedName: 'Dogo Argentino',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 104,
-    breedName: 'Dogo Canario',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 105,
-    breedName: 'Dogo de Burdeos',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 106,
-    breedName: 'Dogo del Tíbet',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 107,
-    breedName: 'Dogo Español',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 108,
-    breedName: 'Elkhound Noruego Gris',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 109,
-    breedName: 'Elkhound Noruego Negro',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 110,
-    breedName: 'English Springer Spaniel',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 111,
-    breedName: 'English Toy Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 112,
-    breedName: 'Épagneul Breton',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 113,
-    breedName: 'Eurasier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 114,
-    breedName: 'Field Spaniel',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 115,
-    breedName: 'Fila Brasileiro',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 116,
-    breedName: 'Fila de San Miguel',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 117,
-    breedName: 'Flat-Coated Retriever',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 118,
-    breedName: 'Fox Terrier de Pelo Duro',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 119,
-    breedName: 'Fox Terrier de Pelo Liso',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 120,
-    breedName: 'Fox Terrier Toy',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 121,
-    breedName: 'Foxhound Americano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 122,
-    breedName: 'Foxhound Inglés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 123,
-    breedName: 'Galgo Afgano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 124,
-    breedName: 'Galgo Español',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 125,
-    breedName: 'Golden Retriever',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 126,
-    breedName: 'Gonzcy Polski',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 127,
-    breedName: 'Gos d Atura Català - Perro de Pastor Catalán',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 128,
-    breedName: 'Gos Rater Valencià',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 129,
-    breedName: 'Gran Azul de Gascuña',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 130,
-    breedName: 'Gran Boyero Suizo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 131,
-    breedName: 'Gran Grifón Vendeano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 132,
-    breedName: 'Gran Münsterlander',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 133,
-    breedName: 'Gran Sabueso Anglo-Francés Blanco y Naranja',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 134,
-    breedName: 'Gran Sabueso Anglo-Francés Blanco y Negro',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 135,
-    breedName: 'Gran Sabueso Anglo-Francés Tricolor',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 136,
-    breedName: 'Gran Spitz - Grossspitz',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 137,
-    breedName: 'Grand Basset Griffon Vendéen',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 138,
-    breedName: 'Grand Gascon Saintongeois',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 139,
-    breedName: 'Greyhound',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 140,
-    breedName: 'Griffon À Poil Laineux',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 141,
-    breedName: 'Griffon Fauve de Bretagne',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 142,
-    breedName: 'Griffon Nivernais',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 143,
-    breedName: 'Grifón Azul de Gascuña',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 144,
-    breedName: 'Grifón Belga',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 145,
-    breedName: 'Grifón Bohemio de Pelo Duro',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 146,
-    breedName: 'Grifón de Bruselas',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 147,
-    breedName: 'Grifón de Muestra Eslovaco de Pelo Duro',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 148,
-    breedName: 'Grifón de Muestra Korthals de Pelo Duro',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 149,
-    breedName: 'Harrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 150,
-    breedName: 'Hokkaïdo Inu',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 151,
-    breedName: 'Hollandse Smoushond',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 152,
-    breedName: 'Hovawart',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 153,
-    breedName: 'Husky Siberiano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 154,
-    breedName: 'Irish Glen of Imaal Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 155,
-    breedName: 'Irish Soft Coated Wheaten Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 156,
-    breedName: 'Irish Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 157,
-    breedName: 'Jack Russell Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 158,
-    breedName: 'Jämthund',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 159,
-    breedName: 'Kai',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 160,
-    breedName: 'Kelpie',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 161,
-    breedName: 'Kerry Blue Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 162,
-    breedName: 'King Charles Spaniel',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 163,
-    breedName: 'Kishu',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 164,
-    breedName: 'Komondor',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 165,
-    breedName: 'Kromfohrländer',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 166,
-    breedName: 'Kuvasz',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 167,
-    breedName: 'Kyi-Leo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 168,
-    breedName: 'Labrador Retriever',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 169,
-    breedName: 'Laika de Siberia Occidental',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 170,
-    breedName: 'Laika de Siberia Oriental',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 171,
-    breedName: 'Laika Ruso-Europeo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 172,
-    breedName: 'Lakeland Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 173,
-    breedName: 'Lancashire Heeler',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 174,
-    breedName: 'Landseer',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 175,
-    breedName: 'Lebrel Húngaro, ',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 176,
-    breedName: 'Lebrel Irlandés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 177,
-    breedName: 'Lebrel Polonés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 178,
-    breedName: 'Leonberger',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 179,
-    breedName: 'Lhasa Apso',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 180,
-    breedName: 'Manchester Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 181,
-    breedName: 'Mastiff',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 182,
-    breedName: 'Mastín del Pirineo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 183,
-    breedName: 'Mastín Español',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 184,
-    breedName: 'Mastín Napolitano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 185,
-    breedName: 'Mestizos gigantes - criollo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 186,
-    breedName: 'Mestizos grandes - criollo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 187,
-    breedName: 'Mestizos medianos - criollo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 188,
-    breedName: 'Mestizos pequeños - criollo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 189,
-    breedName: 'Mudi',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 190,
-    breedName: 'Norfolk Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 191,
-    breedName: 'Norsk Buhund',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 192,
-    breedName: 'Norsk Lundehund',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 193,
-    breedName: 'Norwich Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 194,
-    breedName: 'Nova Scotia Duck Tolling Retriever',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 195,
-    breedName: 'Papillon',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 196,
-    breedName: 'Parson Russell Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 197,
-    breedName: 'Pastor Alemán',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 198,
-    breedName: 'Pastor Australiano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 199,
-    breedName: 'Pastor Australiano Miniatura',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 200,
-    breedName: 'Pastor Australiano Stumpy Tail',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 201,
-    breedName: 'Pastor Belga Groenendael',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 202,
-    breedName: 'Pastor Belga Laekenois',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 203,
-    breedName: 'Pastor Belga Malinois',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 204,
-    breedName: 'Pastor Belga Tervueren',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 205,
-    breedName: 'Pastor Blanco Suizo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 206,
-    breedName: 'Pastor Croato',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 207,
-    breedName: 'Pastor de Anatolia',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 208,
-    breedName: 'Pastor de Asia Central',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 209,
-    breedName: 'Pastor de Beauce',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 210,
-    breedName: 'Pastor de Bérgamo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 211,
-    breedName: 'Pastor de Bosnia-Herzegovina y Croacia',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 212,
-    breedName: 'Pastor de Brie',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 213,
-    breedName: 'Pastor de Karst',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 214,
-    breedName: 'Pastor de los Pirineos de cara rasa',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 215,
-    breedName: 'Pastor de los Pirineos de pelo largo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 216,
-    breedName: 'Pastor de Picardía',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 217,
-    breedName: 'Pastor de Rusia Meridional',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 218,
-    breedName: 'Pastor del Cáucaso',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 219,
-    breedName: 'Pastor Finlandés de Laponia',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 220,
-    breedName: 'Pastor Garafiano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 221,
-    breedName: 'Pastor Holandés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 222,
-    breedName: 'Pastor Islandés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 223,
-    breedName: 'Pastor Maremmano-Abruzzese',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 224,
-    breedName: 'Pastor Polaco de las Llanuras',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 225,
-    breedName: 'Pastor Polaco de Podhale',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 226,
-    breedName: 'Pastor Vasco del Gorbea - Euskal Artzain Txakurra Gorbeiakoa',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 227,
-    breedName: 'Pastor Yugoslavo de Charplanina',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 228,
-    breedName: 'Pekinés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 229,
-    breedName: 'Pequeño Azul de Gascuña',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 230,
-    breedName: 'Pequeño Lebrel Italiano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 231,
-    breedName: 'Pequeño Münsterländer',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 232,
-    breedName: 'Pequeño Perro Holandes para la Caza Acuática',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 233,
-    breedName: 'Pequeño Perro Ruso',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 234,
-    breedName: 'Perdiguero Alemán',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 235,
-    breedName: 'Perdiguero de Burgos',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 236,
-    breedName: 'Perdiguero de Drente',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 237,
-    breedName: 'Perdiguero Portugués',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 238,
-    breedName: 'Perro Crestado Chino',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 239,
-    breedName: 'Perro de Agua Americano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 240,
-    breedName: 'Perro de Agua Español',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 241,
-    breedName: 'Perro de Agua Frisón',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 242,
-    breedName: 'Perro de Agua Irlandés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 243,
-    breedName: 'Perro de Agua Portugués',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 244,
-    breedName: 'Perro de Groenlandia',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 245,
-    breedName: 'Perro de Montaña de los Pirineos',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 246,
-    breedName: 'Perro de Muestra Alemán de Pelo Cerdoso',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 247,
-    breedName: 'Perro de Muestra Alemán de Pelo Largo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 248,
-    breedName: 'Perro de Muestra Frisón',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 249,
-    breedName: 'Perro de Nutria',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 250,
-    breedName: 'Perro de Osos de Carelia',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 251,
-    breedName: 'Perro de Pastor Rumano de los Carpatos ""Carpatin""',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 252,
-    breedName: 'Perro de Pastor Rumano de Mioritza',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 253,
-    breedName: 'Perro de San Huberto',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 254,
-    breedName: 'Perro de Taiwan',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 255,
-    breedName: 'Perro del Faraón',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 256,
-    breedName: 'Perro Finlandés de Laponia',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 257,
-    breedName: 'Perro Jindo Coreano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 258,
-    breedName: 'Perro Lobo Checoslovaco',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 259,
-    breedName: 'Perro Lobo de Saarloos',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 260,
-    breedName: 'Perro Majorero',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 261,
-    breedName: 'Perro Pastor de las Shetland',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 262,
-    breedName: 'Perro Sin Pelo de Perú',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 263,
-    breedName: 'Perro Sueco de Laponia',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 264,
-    breedName: 'Perro Tejonero Alpino',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 265,
-    breedName: 'Perro Tejonero de Westfalia',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 266,
-    breedName: 'Perro Tejonero Sueco',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 267,
-    breedName: 'Petit Basset Griffon Vendéen',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 268,
-    breedName: 'Petit Brabançon',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 269,
-    breedName: 'Petit Chien Lion',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 270,
-    breedName: 'Petit Gascon Saintongeois',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 271,
-    breedName: 'Pinscher Alemán',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 272,
-    breedName: 'Pinscher Austríaco de Pelo Corto',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 273,
-    breedName: 'Pinscher Miniatura',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 274,
-    breedName: 'Plott Hound',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 275,
-    breedName: 'Podenco Andaluz',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 276,
-    breedName: 'Podenco Canario',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 277,
-    breedName: 'Podenco Ibicenco',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 278,
-    breedName: 'Podenco Portugués de Pelo Áspero',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 279,
-    breedName: 'Podenco Portugués de Pelo Liso',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 280,
-    breedName: 'Pointer',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 281,
-    breedName: 'Poitevin',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 282,
-    breedName: 'Pomerania',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 283,
-    breedName: 'Porcelaine',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 284,
-    breedName: 'Pudelpointer',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 285,
-    breedName: 'Puli',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 286,
-    breedName: 'Pumi',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 287,
-    breedName: 'Rafeiro do Alentejo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 288,
-    breedName: 'Ratón de Praga',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 289,
-    breedName: 'Ratonero Bodeguero Andaluz',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 290,
-    breedName: 'Redbone Coonhound',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 291,
-    breedName: 'Retriever de Pelo Rizado',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 292,
-    breedName: 'Rhodesian Ridgeback',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 293,
-    breedName: 'Rottweiler',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 294,
-    breedName: 'Sabueso Alemán',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 295,
-    breedName: 'Sabueso Austriaco Negro y Fuego',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 296,
-    breedName: 'Sabueso de Bosnia de Pelo Cerdoso',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 297,
-    breedName: 'Sabueso de Hamilton',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 298,
-    breedName: 'Sabueso de Hygen',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 299,
-    breedName: 'Sabueso de Istria de Pelo Corto',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 300,
-    breedName: 'Sabueso de Istria de Pelo Duro',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 301,
-    breedName: 'Sabueso de los Balcanes',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 302,
-    breedName: 'Sabueso de Montaña del Montenegro',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 303,
-    breedName: 'Sabueso de Posavaz',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 304,
-    breedName: 'Sabueso de Sangre de Baviera',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 305,
-    breedName: 'Sabueso de Sangre de Hanover',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 306,
-    breedName: 'Sabueso de Schiller',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 307,
-    breedName: 'Sabueso de Småland',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 308,
-    breedName: 'Sabueso de Tirol',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 309,
-    breedName: 'Sabueso de Transilvania',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 310,
-    breedName: 'Sabueso del Ariège',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 311,
-    breedName: 'Sabueso Eslovaco',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 312,
-    breedName: 'Sabueso Español',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 313,
-    breedName: 'Sabueso Estirio de Pelo Duro',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 314,
-    breedName: 'Sabueso Finlandés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 315,
-    breedName: 'Sabueso fino colombiano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 316,
-    breedName: 'Sabueso Francés Blanco y Naranja',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 317,
-    breedName: 'Sabueso Francés Blanco y Negro',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 318,
-    breedName: 'Sabueso Francés Tricolor',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 319,
-    breedName: 'Sabueso Halden',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 320,
-    breedName: 'Sabueso Helénico',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 321,
-    breedName: 'Sabueso Italiano de Pelo Corto',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 322,
-    breedName: 'Sabueso Italiano de Pelo Duro',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 323,
-    breedName: 'Sabueso Noruego',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 324,
-    breedName: 'Sabueso Polaco',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 325,
-    breedName: 'Sabueso Serbio',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 326,
-    breedName: 'Sabueso Suizo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 327,
-    breedName: 'Sabueso Suizo Pequeño',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 328,
-    breedName: 'Sabueso Tricolor Serbio',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 329,
-    breedName: 'Saluki',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 330,
-    breedName: 'Samoyedo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 331,
-    breedName: 'San Bernardo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 332,
-    breedName: 'Schapendoes',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 333,
-    breedName: 'Schipperke',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 334,
-    breedName: 'Schnauzer Gigante',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 335,
-    breedName: 'Schnauzer Mediano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 336,
-    breedName: 'Schnauzer Miniatura',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 337,
-    breedName: 'Scottish Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 338,
-    breedName: 'Sealyham Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 339,
-    breedName: 'Setter Gordon',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 340,
-    breedName: 'Setter Inglés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 341,
-    breedName: 'Setter Irlandés Rojo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 342,
-    breedName: 'Setter Irlandés Rojo y Blanco',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 343,
-    breedName: 'Shar Pei',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 344,
-    breedName: 'Shiba Inu',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 345,
-    breedName: 'Shih Tzu',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 346,
-    breedName: 'Shikoku Inu',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 347,
-    breedName: 'Silky Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 348,
-    breedName: 'Skye Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 349,
-    breedName: 'Sloughi',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 350,
-    breedName: 'Spaniel Azul de Picardía',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 351,
-    breedName: 'Spaniel de Pont-Audemer',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 352,
-    breedName: 'Spaniel Francés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 353,
-    breedName: 'Spaniel Japonés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 354,
-    breedName: 'Spaniel Picardo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 355,
-    breedName: 'Spaniel Tibetano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 356,
-    breedName: 'Spinone Italiano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 357,
-    breedName: 'Spitz de Norrbotten',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 358,
-    breedName: 'Spitz Finlandés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 359,
-    breedName: 'Spitz Japonés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 360,
-    breedName: 'Spitz Lobo - Wolfsspitz',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 361,
-    breedName: 'Spitz mediano - Mittelspitz',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 362,
-    breedName: 'Spitz pequeño - Kleinspitz',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 363,
-    breedName: 'Staffordshire Bull Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 364,
-    breedName: 'Suomenajokoira',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 365,
-    breedName: 'Sussex Spaniel',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 366,
-    breedName: 'Tchuvatch Eslovaco',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 367,
-    breedName: 'Teckel',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 368,
-    breedName: 'Terranova',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 369,
-    breedName: 'Terrier Alemán de Caza',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 370,
-    breedName: 'Terrier Brasileiro',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 371,
-    breedName: 'Terrier de Norfolk',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 372,
-    breedName: 'Terrier Japonés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 373,
-    breedName: 'Terrier Negro Ruso',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 374,
-    breedName: 'Terrier Tibetano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 375,
-    breedName: 'Thai Ridgeback Dog',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 376,
-    breedName: 'Tosa',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 377,
-    breedName: 'Treeing Walker Coonhound',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 378,
-    breedName: 'Vallhund Sueco',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 379,
-    breedName: 'Volpino Italiano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 380,
-    breedName: 'Weimaraner',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 381,
-    breedName: 'Welsh Corgi Cardigan',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 382,
-    breedName: 'Welsh Corgi Pembroke',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 383,
-    breedName: 'Welsh Springer Spaniel',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 384,
-    breedName: 'Welsh Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 385,
-    breedName: 'West Highland White Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 386,
-    breedName: 'Whippet',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 387,
-    breedName: 'Xoloitzcuintle - Perro Sin Pelo Mexicano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 388,
-    breedName: 'Yorkshire Terrier',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 389,
-    breedName: 'Gatos Mestizo de pelo corto',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 390,
-    breedName: 'Gatos Mestizo de pelo semilargo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 391,
-    breedName: 'Gatos Mestizo de pelo largo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 392,
-    breedName: 'Abisinio',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 393,
-    breedName: 'Amercian Curl de pelo largo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 394,
-    breedName: 'American Bobtail',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 395,
-    breedName: 'American Curl',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 396,
-    breedName: 'American Shorthair',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 397,
-    breedName: 'American Wirehair',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 398,
-    breedName: 'Anatoli',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 399,
-    breedName: 'Angora Turco',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 400,
-    breedName: 'Asian',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 401,
-    breedName: 'Australian Mist',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 402,
-    breedName: 'Azul Ruso',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 403,
-    breedName: 'Balinés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 404,
-    breedName: 'Bengalí',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 405,
-    breedName: 'Bombay',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 406,
-    breedName: 'Bosque de Noruega',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 407,
-    breedName: 'Brasileño de Pelo Corto',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 408,
-    breedName: 'Británico de pelo corto',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 409,
-    breedName: 'Británico de Pelo Corto Azul',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 410,
-    breedName: 'Británico de pelo largo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 411,
-    breedName: 'Burmés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 412,
-    breedName: 'Burmilla',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 413,
-    breedName: 'California Spangled',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 414,
-    breedName: 'Cartujo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 415,
-    breedName: 'Ceilán',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 416,
-    breedName: 'Celta de Pelo Corto',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 417,
-    breedName: 'Cornish Rex',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 418,
-    breedName: 'Devon Rex',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 419,
-    breedName: 'Don Sphynx',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 420,
-    breedName: 'Europeo de Pelo Corto',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 421,
-    breedName: 'Exótico de pelo corto',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 422,
-    breedName: 'Exótico de pelo largo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 423,
-    breedName: 'Foldex',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 424,
-    breedName: 'German Rex',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 425,
-    breedName: 'Habana Brown',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 426,
-    breedName: 'Highland Fold',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 427,
-    breedName: 'Highland Straight',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 428,
-    breedName: 'Himalayo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 429,
-    breedName: 'Javanés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 430,
-    breedName: 'Kannani',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 431,
-    breedName: 'Korat',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 432,
-    breedName: 'Kurilean Bobtail de Pelo Corto',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 433,
-    breedName: 'Kurilean Bobtail de Pelo Largo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 434,
-    breedName: 'Laperm de Pelo Corto',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 435,
-    breedName: 'Laperm de Pelo Largo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 436,
-    breedName: 'Maine Coon',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 437,
-    breedName: 'Manx',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 438,
-    breedName: 'Mau Egipcio',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 439,
-    breedName: 'Munchkin',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 440,
-    breedName: 'Nebelung',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 441,
-    breedName: 'Neva Masquerade',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 442,
-    breedName: 'Ocicat',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 443,
-    breedName: 'Oriental de Pelo Corto',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 444,
-    breedName: 'Oriental de Pelo Semilargo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 445,
-    breedName: 'Persa',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 446,
-    breedName: 'Peterbald',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 447,
-    breedName: 'Pixiebob',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 448,
-    breedName: 'Ragdoll',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 449,
-    breedName: 'Sagrado de Birmania',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 450,
-    breedName: 'Scottish Fold',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 451,
-    breedName: 'Selkirk Rex',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 452,
-    breedName: 'Seychellois de Pelo Corto',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 453,
-    breedName: 'Seychellois de Pelo Largo',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 454,
-    breedName: 'Siamés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 455,
-    breedName: 'Siberiano',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 456,
-    breedName: 'Singapura',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 457,
-    breedName: 'Snowshoe',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 458,
-    breedName: 'Sokoke',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 459,
-    breedName: 'Somalí',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 460,
-    breedName: 'Sphynx',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 461,
-    breedName: 'Thai',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 462,
-    breedName: 'Tonkinés',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 463,
-    breedName: 'Toyger',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 464,
-    breedName: 'Van Turco',
-    specieId: 2,
-    specieName: 'Perro',
-  ),
-  BreedModel(
-    breedId: 465,
-    breedName: 'York',
-    specieId: 2,
-    specieName: 'Perro',
-  )
-];
