--- lib/config/utils/data/deworming.list.dart
@@ -1,7 +1,0 @@
-import 'package:petmeals/config/utils/model/deworming.model.dart';
-
-final dewormingList = [
-  DewormingModel(typeId: 1, typeName: 'Interna'),
-  DewormingModel(typeId: 2, typeName: 'Externa'),
-  DewormingModel(typeId: 3, typeName: 'Interna - Externa')
-];
