--- lib/config/utils/data/gender.list.dart
@@ -1,7 +1,0 @@
-import 'package:petmeals/config/utils/model/gender.model.dart';
-
-final genderList = [
-  GenderModel(value: 0, name: 'Mujer'),
-  GenderModel(value: 1, name: 'Hombre'),
-  GenderModel(value: 2, name: 'No binario'),
-];
