--- lib/config/utils/data/grooming.list.dart
@@ -1,14 +1,0 @@
-import 'package:petmeals/config/utils/model/grooming.model.dart';
-
-final groomingList = [
-  GroomingModel(typeId: 1, typeName: 'Baño'),
-  GroomingModel(typeId: 2, typeName: 'Corte'),
-  GroomingModel(typeId: 3, typeName: 'Baño y Corte'),
-  GroomingModel(typeId: 4, typeName: 'Baño Medicado'),
-  GroomingModel(typeId: 5, typeName: 'Corte de uñas'),
-  GroomingModel(typeId: 6, typeName: 'Limpieza de oído'),
-  GroomingModel(typeId: 7, typeName: 'Limpieza de glándulas anales'),
-  GroomingModel(typeId: 8, typeName: 'Limpieza dental'),
-  GroomingModel(typeId: 9, typeName: 'Cepillado'),
-  GroomingModel(typeId: 10, typeName: 'Perfumado')
-];
