--- lib/config/utils/data/vaccine.list.dart
@@ -1,22 +1,0 @@
-import 'package:petmeals/config/utils/model/vaccine.model.dart';
-
-final vaccineList = [
-  VaccineModel(typeId: 1, typeName: 'Leptospira', especieId: 2),
-  VaccineModel(typeId: 2, typeName: 'Distemper', especieId: 2),
-  VaccineModel(typeId: 3, typeName: 'Parvorirus', especieId: 2),
-  VaccineModel(typeId: 4, typeName: 'Antirrabica', especieId: 2),
-  VaccineModel(typeId: 5, typeName: 'Triple', especieId: 2),
-  VaccineModel(typeId: 6, typeName: 'Cuadruple', especieId: 2),
-  VaccineModel(typeId: 7, typeName: 'Quintuple', especieId: 2),
-  VaccineModel(typeId: 8, typeName: 'Quintuple + Rabia', especieId: 2),
-  VaccineModel(typeId: 9, typeName: 'Sextuple', especieId: 2),
-  VaccineModel(typeId: 10, typeName: 'Sextuple + Rabia', especieId: 2),
-  VaccineModel(typeId: 11, typeName: 'Leucemia', especieId: 2),
-  VaccineModel(typeId: 12, typeName: 'Tos de las perreras', especieId: 2),
-  VaccineModel(typeId: 13, typeName: 'Refuerzo anual', especieId: 2),
-  VaccineModel(typeId: 14, typeName: 'Triple', especieId: 1),
-  VaccineModel(typeId: 15, typeName: 'Triple + Rabia', especieId: 1),
-  VaccineModel(typeId: 16, typeName: 'Leucemia', especieId: 1),
-  VaccineModel(typeId: 17, typeName: 'Antirrabica', especieId: 1),
-  VaccineModel(typeId: 18, typeName: 'Refuerzo anual', especieId: 1)
-];
