--- lib/config/utils/model/breed.model.dart
@@ -1,13 +1,0 @@
-class BreedModel {
-  BreedModel({
-    this.breedId,
-    this.breedName,
-    this.specieId,
-    this.specieName,
-  });
-
-  final int? breedId;
-  final String? breedName;
-  final int? specieId;
-  final String? specieName;
-}
