--- lib/config/utils/model/deworming.model.dart
@@ -1,9 +1,0 @@
-class DewormingModel {
-  DewormingModel({
-    this.typeId,
-    this.typeName,
-  });
-
-  final int? typeId;
-  final String? typeName;
-}
