--- lib/config/utils/model/gender.model.dart
@@ -1,9 +1,0 @@
-class GenderModel {
-  GenderModel({
-    this.value,
-    this.name,
-  });
-
-  final int? value;
-  final String? name;
-}
