--- lib/config/utils/model/grooming.model.dart
@@ -1,9 +1,0 @@
-class GroomingModel {
-  GroomingModel({
-    this.typeId,
-    this.typeName,
-  });
-
-  final int? typeId;
-  final String? typeName;
-}
