--- lib/config/utils/model/vaccine.model.dart
@@ -1,11 +1,0 @@
-class VaccineModel {
-  VaccineModel({
-    this.typeId,
-    this.typeName,
-    this.especieId,
-  });
-
-  final int? typeId;
-  final String? typeName;
-  final int? especieId;
-}
