+++ lib/main.dart
@@ -1,5 +1,5 @@
-import 'package:petmeals/app/presentation/main.get_it.dart';
-import 'package:petmeals/app/presentation/setup.get_it.dart';
+import 'package:petmeals/main.get_it.dart';
+import 'package:petmeals/setup.get_it.dart';
 import 'package:petmeals/firebase_options.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
